Rename Settings -> FindBugsKeys
