+++ src/main/scala/de/johoop/findbugs4sbt/Filters.scala
@@ -17,7 +17,7 @@ import sbt._
 import scala.xml.Node
 import scala.xml.XML
 
-private[findbugs4sbt] trait Filters extends Plugin with Settings {
+private[findbugs4sbt] trait Filters extends Plugin with FindBugsKeys {
 
   private[findbugs4sbt] def addFilterFiles(filters: FilterSettings, filterPath: File, options: List[String]) = {
     def addIncludeFilterFile(options: List[String]) = addFilterFile(options, filters.includeFilters, "include")
