+++ src/main/scala/de/johoop/findbugs4sbt/FindBugs.scala
@@ -15,9 +15,9 @@ package de.johoop.findbugs4sbt
 import sbt._
 import Keys._
 
-object FindBugs extends Plugin with Settings with CommandLine with CommandLineExecutor {
+object FindBugs extends Plugin with FindBugsKeys with CommandLine with CommandLineExecutor {
 
-  override def findbugsTask(
+  def findbugsTask(
       findbugsClasspath: Classpath,
       compileClasspath: Classpath,
       paths: PathSettings,
@@ -32,4 +32,48 @@ object FindBugs extends Plugin with Settings with CommandLine with CommandLineEx
       executeCommandLine(cmd, javaHome, streams.log)
     }
   }
+
+  val findbugsSettings = Seq(
+    ivyConfigurations += FindbugsConfig,
+    libraryDependencies ++= Seq(
+      "com.google.code.findbugs" % "findbugs" % "3.0.1" % "findbugs->default",
+      "com.google.code.findbugs" % "jsr305" % "3.0.1" % "findbugs->default"
+    ),
+    findbugs := findbugsTask(
+      findbugsClasspath.value,
+      (managedClasspath in Compile).value,
+      findbugsPathSettings.value,
+      findbugsFilterSettings.value,
+      findbugsMiscSettings.value,
+      javaHome.value,
+      streams.value
+    ),
+    findbugsPathSettings := PathSettings(
+      findbugsReportPath.value,
+      findbugsAnalyzedPath.value,
+      findbugsAuxiliaryPath.value), //.dependsOn(compile in Compile),
+    findbugsFilterSettings := FilterSettings(findbugsIncludeFilters.value, findbugsExcludeFilters.value),
+    findbugsMiscSettings := MiscSettings(
+      findbugsReportType.value,
+      findbugsPriority.value,
+      findbugsOnlyAnalyze.value,
+      findbugsMaxMemory.value,
+      findbugsAnalyzeNestedArchives.value,
+      findbugsSortReportByClassNames.value,
+      findbugsEffort.value
+    ),
+    findbugsClasspath := Classpaths.managedJars(FindbugsConfig, classpathTypes.value, update.value),
+    findbugsReportType := Some(ReportType.Xml),
+    findbugsPriority := Priority.Medium,
+    findbugsEffort := Effort.Default,
+    findbugsReportPath := Some(crossTarget.value / "findbugs" / "report.xml"),
+    findbugsMaxMemory := 1024,
+    findbugsAnalyzeNestedArchives := true,
+    findbugsSortReportByClassNames := false,
+    findbugsAnalyzedPath := Seq((classDirectory in Compile).value),
+    findbugsAuxiliaryPath := (dependencyClasspath in Compile).value.files,
+    findbugsOnlyAnalyze := None,
+    findbugsIncludeFilters := None,
+    findbugsExcludeFilters := None
+  )
 }
